+++ crates/vmi-utils/src/injector/os/windows/user_mode.rs
@@ -115,11 +115,14 @@ where
     fn with_bridge<Bridge>(
         self,
         bridge: Bridge,
-    ) -> <WindowsOs<Driver> as InjectorExecutionAdapter<UserMode, T, Bridge>>::Handler
+    ) -> Result<
+        <WindowsOs<Driver> as InjectorExecutionAdapter<UserMode, T, Bridge>>::Handler,
+        VmiError,
+    >
     where
         Bridge: BridgeDispatch<WindowsOs<Driver>>,
     {
-        UserInjectorHandler {
+        Ok(UserInjectorHandler {
             pid: self.pid,
             tid: self.tid,
             view: self.view,
@@ -129,7 +132,7 @@ where
             bridge,
             state: InjectorState::PreHijack,
             output: None,
-        }
+        })
     }
 }